@@ -15,7 +15,7 @@
     "dev:local:ddev": ". ./test/.env.ddev && npm run dev:local",
     "prod:local:ddev": ". ./test/.env.ddev.prod && docker-compose rm -rf && docker-compose up",
     "lint": "eslint .",
-    "postinstall": "npm run build -- --colors",
+    "postinstall": "npm run build -- --color=true",
     "start": "node .",
     "test": ". ./test/.env.local && mocha"
   },
@@ -32,52 +32,53 @@
   "homepage": "https://bitbucket.org/atlassian/a11y-dashboard#readme",
   "dependencies": {
     "bunyan": "^1.5.1",
-    "compression": "^1.6.0",
-    "express": "^4.13.3"
+    "compression": "^1.6.1",
+    "express": "^4.13.4"
   },
   "devDependencies": {
-    "babel-core": "^6.3.17",
-    "babel-loader": "^6.2.0",
+    "babel-core": "^6.4.5",
+    "babel-loader": "^6.2.1",
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.3.13",
     "chai": "^3.4.1",
-    "chai-as-promised": "^5.1.0",
+    "chai-as-promised": "^5.2.0",
     "css-loader": "^0.23.1",
     "debug": "^2.2.0",
     "deep-equal": "^1.0.1",
     "es6-promise": "^3.0.2",
     "eslint": "^1.10.3",
-    "eslint-config-airbnb": "^2.0.0",
-    "eslint-plugin-react": "^3.13.1",
+    "eslint-config-airbnb": "^4.0.0",
+    "eslint-plugin-react": "^3.16.1",
     "exports-loader": "^0.6.2",
     "finchjs": "^0.5.14",
-    "highlight.js": "^9.0.0",
-    "history": "^1.17.0",
+    "form-serialize": "^0.7.0",
+    "highlight.js": "^9.1.0",
+    "history": "^2.0.0-rc2",
     "html-loader": "^0.4.0",
-    "html-webpack-plugin": "^1.7.0",
+    "html-webpack-plugin": "^2.7.2",
     "json-loader": "^0.5.4",
     "less": "^2.5.3",
     "less-loader": "^2.2.2",
-    "lodash": "^3.10.1",
-    "mocha": "^2.3.4",
+    "lodash": "^4.0.1",
+    "mocha": "^2.4.2",
     "mocha-junit-reporter": "^1.9.1",
-    "moment": "^2.10.6",
+    "moment": "^2.11.1",
     "mustache": "^2.2.1",
     "nodemon": "^1.8.1",
     "object-assign": "^4.0.1",
     "parallelshell": "^2.0.0",
     "query-string": "^3.0.0",
-    "react": "^0.14.5",
-    "react-dom": "^0.14.5",
+    "react": "^0.14.6",
+    "react-dom": "^0.14.6",
     "react-google-charts": "^0.2.0",
-    "react-router": "^1.0.3",
-    "rimraf": "^2.4.4",
+    "react-router": "^2.0.0-rc5",
+    "rimraf": "^2.5.1",
     "style-loader": "^0.13.0",
     "supertest": "^1.1.0",
-    "wait-on": "^1.2.1",
-    "webpack": "^1.12.9",
-    "webpack-dev-middleware": "^1.4.0",
-    "webpack-hot-middleware": "^2.6.0",
+    "wait-on": "^1.3.1",
+    "webpack": "^2.0.5-beta",
+    "webpack-dev-middleware": "^1.5.1",
+    "webpack-hot-middleware": "^2.6.4",
     "xhr": "^2.2.0"
   }
 }
