@@ -1,37 +1,8 @@
 import React from 'react';
 import levels from '../data/levels.json';
 import logger from '../logger';
-import queryString from 'query-string';
-import history from '../history';
-import objectAssign from 'object-assign';
-import _ from 'lodash';
-
-function historyUpdate(newState) {
-  setTimeout(() => {
-    history.replace({
-      pathname: window.location.pathname,
-      search: '?' + queryString.stringify(objectAssign({}, queryString.parse(window.location.search), newState)),
-    });
-  });
-}
-const changeLevel = (e) => {
-  logger.debug(`Changed level to ${e.target.value}`);
-  historyUpdate({ level: e.target.value || 'error' });
-};
-
-function initLevelSelect() {
-  logger.debug('init level select');
-  const levelSelect = document.getElementById('level');
-  if (levelSelect) {
-    levelSelect.removeEventListener('change', changeLevel);
-    levelSelect.addEventListener('change', changeLevel);
-  }
-}
-
-const urlUpdate = _.debounce((url) => {
-  logger.debug('filter URLs by', url);
-  historyUpdate({ reverseDns: url || undefined });
-}, 250);
+import { browserHistory } from 'react-router';
+import serialize from 'form-serialize';
 
 class Filter extends React.Component {
 
@@ -42,24 +13,19 @@ class Filter extends React.Component {
     };
   }
 
-    componentDidMount() {
-      initLevelSelect();
-    }
-
-    componentDidUpdate() {
-      initLevelSelect();
-    }
-
-    onClearClick(e) {
+    onFormSubmit(e) {
       e.preventDefault();
-      this.setState({
-        reverseDns: '',
+      const query = serialize(e.target);
+      browserHistory.push({
+        pathname: window.location.pathname,
+        search: `?${query}`,
       });
-      historyUpdate({ reverseDns: null });
     }
 
     render() {
-      let selectElement = null;
+      logger.debug('rendering filter');
+
+      let levelSelect = null;
       if (this.props.levels.length) {
         const options = this.props.levels.map((level) => {
           const extraAttrs = {};
@@ -69,74 +35,74 @@ class Filter extends React.Component {
           return <aui-option key={level} {...extraAttrs}>{level}</aui-option>;
         });
 
-        selectElement = (
-          <span>
-            <aui-label for="level">Level</aui-label>
-            <aui-select value={this.props.currentLevel} id="level" name="level" placeholder="Select a level">{options}</aui-select>
-          </span>
+        levelSelect = (
+          <div className="field-group">
+            <aui-label for="level">Error level</aui-label>
+            <aui-select
+              class="full-width-field"
+              value={this.props.currentLevel}
+              id="level"
+              name="level"
+              placeholder="Select a level"
+            >{options}</aui-select>
+          </div>
         );
       }
 
       let standardsSelect = null;
       if (this.props.standards.length) {
-        const initStandardsSelect = (select) => {
-          logger.debug('init standards select');
-          AJS.$(select)
-            .auiSelect2()
-            .off('change')
-            .on('change', (e) => {
-              logger.debug(`Changed standards to ${e.val}`);
-              historyUpdate({ standards: e.val });
-            });
-        };
-
-        const options = this.props.standards.map((standard) => {
-          return <option key={standard} value={standard || 'best-practice'}>{standard || 'best practice'}</option>;
+        const checkboxes = this.props.standards.map((standard) => {
+          const key = standard || 'best-practice';
+          const value = standard || 'best practice';
+          const checked = this.props.selectedStandards.indexOf(standard) !== -1;
+          return (<div key={key} className="checkbox">
+                    <input
+                      className="checkbox"
+                      type="checkbox"
+                      name="standard"
+                      defaultChecked={checked}
+                      value={key}
+                      id={key}
+                    />
+                    <label htmlFor={key}>{value}</label>
+                  </div>);
         });
 
-        standardsSelect = (<span>
-                            <label htmlFor="standards">Standard</label>
-                            <select id="standards" multiple="multiple" ref={initStandardsSelect} value={this.props.selectedStandards} onChange={() => {}}>
-                                {options}
-                            </select>
-                          </span>);
-      }
-
-      const value = this.state.reverseDns === null ? this.props.reverseDns : this.state.reverseDns;
-
-      const onUrlChange = (e) => {
-        urlUpdate(e.target.value);
-        this.setState({
-          reverseDns: e.target.value || '',
-        });
-      };
-
-      let reset = null;
-      if (value) {
-        reset = (<a className="cancel" href="#reset" onClick={this.onClearClick.bind(this)}>Reset</a>);
+        standardsSelect = (
+            <fieldset className="group">
+                 <legend><span>Tags</span></legend>
+                 {checkboxes}
+             </fieldset>
+        );
       }
 
       return (
-      <form className="aui">
-          <div className="aui-group">
-              <div className="aui-item">
-                  <strong>Filter:
-                  </strong>
-                  <label htmlFor="url" title="reverse domain name notation">Reverse DNS</label>
-                  <input id="reverseDns" className="text medium-field" name="reverseDns" ref="reverseDns" placeholder="%" type="text" onChange={onUrlChange} value={value}/>
-                  <div className="aui-buttons">
-                      {reset}
-                  </div>
-                  {selectElement}
-                  {standardsSelect}
-              </div>
+        <form className="aui top-label" onSubmit={this.onFormSubmit}>
+          <div className="field-group">
+            <label htmlFor="url" title="Use % as a wildcard">Page URL</label>
+            <input
+              id="url"
+              className="text full-width-field"
+              name="url"
+              placeholder="eg. %jira%admin%"
+              type="text"
+              defaultValue={this.props.url}
+            />
           </div>
-      </form>);
+          {levelSelect}
+          {standardsSelect}
+          <div className="buttons-container">
+            <div className="buttons">
+              <button className="aui-button aui-button-primary submit">Update</button>
+            </div>
+          </div>
+        </form>
+      );
     }
 }
 
 Filter.propTypes = {
-  reverseDns: React.PropTypes.string.isRequired,
+  url: React.PropTypes.string.isRequired,
   standards: React.PropTypes.array.isRequired,
   selectedStandards: React.PropTypes.array.isRequired,
   levels: React.PropTypes.array.isRequired,
