@@ -1,6 +1,7 @@
 import React from 'react';
 import moment from 'moment';
-import _ from 'lodash';
+import sortBy from 'lodash/sortBy';
+import uniq from 'lodash/uniq';
 import queryString from 'query-string';
 import deepEqual from 'deep-equal';
 import objectAssign from 'object-assign';
@@ -16,13 +17,19 @@ require('../styles/details.less');
 
 function transformCulprits(culprits) {
   const data = {};
-  const sortedRows = _.sortByAll(culprits, ['reverse_dns', 'origin_library', 'code', 'message', 'selector']);
+  const sortedRows = sortBy(culprits, [
+    'reverse_dns',
+    'origin_library',
+    'code',
+    'message',
+    'selector',
+  ]);
 
   sortedRows.forEach((row) => {
     data[row.original_url] = data[row.original_url] || {};
     const errorKey = `${row.origin_library}.${row.code}`;
-    data[row.original_url][errorKey] = data[row.original_url][errorKey] || {};
-    data[row.original_url][errorKey][row.message] = data[row.original_url][errorKey][row.message] || {
+    const err = data[row.original_url][errorKey] = data[row.original_url][errorKey] || {};
+    err[row.message] = err[row.message] || {
       rows: [],
       standards: [],
       origin: row.origin_library,
@@ -36,7 +43,7 @@ function transformCulprits(culprits) {
       standards.sort();
     }
     perMessage.rows.push(row);
-    perMessage.rows = _.uniq(perMessage.rows, 'selector');
+    perMessage.rows = uniq(perMessage.rows, 'selector');
   });
   return data;
 }
@@ -68,23 +75,17 @@ class Details extends React.Component {
 
   componentDidMount() {
     this.spinStart();
-    webserviceRequest(`details.stats?${queryString.stringify(this.getBaseQuery(this.props))}`)
-    .then((stats) => {
+    const query = queryString.stringify(this.getBaseQuery(this.props));
+    webserviceRequest(`details.stats?${query}`).then((stats) => {
       logger.debug('stats arrived', stats);
       return stats;
-    })
-    .then((stats) => {
-      this.setState({
-        stats,
-      });
-    })
-    .then(() => this.updateCulprits());
+    }).then((stats) => {
+      this.setState({ stats });
+    }).then(() => this.updateCulprits());
   }
 
   componentWillReceiveProps() {
-    this.setState({
-      culprits: null,
-    });
+    this.setState({ culprits: null });
   }
 
   componentDidUpdate(prevProps) {
@@ -105,9 +106,7 @@ class Details extends React.Component {
     const query = this.getQuery(this.props);
 
     logger.debug('setting query state', query);
-    this.setState({
-      query,
-    });
+    this.setState({ query });
     return query;
   }
 
@@ -118,22 +117,16 @@ class Details extends React.Component {
   getBaseQuery(props) {
     const { project, timestamp } = props.params;
 
-    return {
-      origin: project,
-      timestamp,
-    };
+    return { origin: project, timestamp };
   }
 
   getFilterQuery(props) {
     const { query } = props.location;
-    // by default we will select Section508 and WCAG2AA to show (if available)
-    const queryStandards = query.standards instanceof Array ? query.standards : [query.standards];
-    const selectedStandards = query.standards ? queryStandards : _.intersection(this.state.stats.standards, ['wcag2aa', 'section508']);
 
     return {
-      standard: selectedStandards,
+      standard: query.standard || [],
       level: query.level || 'error',
-      reverseDns: query.reverseDns || undefined,
+      url: query.url || undefined,
     };
   }
 
@@ -142,24 +135,29 @@ class Details extends React.Component {
     this.spinStart();
     const queryParams = this.setQueryState();
 
-    return webserviceRequest(`details?${queryString.stringify(queryParams)}`)
-    .then((culprits) => {
+    return webserviceRequest(`details?${queryString.stringify(queryParams)}`).then((culprits) => {
       logger.debug('transform culprits', culprits);
-      return transformCulprits(culprits);
-    })
-    .then((transformed) => {
+      return {
+        totalFiltered: culprits.length,
+        transformed: transformCulprits(culprits),
+      };
+    }).then(({ transformed, totalFiltered }) => {
       logger.debug('transformed culprits', transformed);
       if (!this.ignoreLastXhr) {
         this.setState({
           culprits: transformed,
+          totalFiltered,
+          urlsFiltered: Object.keys(transformed).length,
         });
       }
     })
-    .then(() => this.spinStop())
-    .catch((err) => logger.error(err));
+      .then(() => this.spinStop())
+      .then(() => AJS.$(this.refs.spinner).hide())
+      .catch((err) => logger.error(err));
   }
 
   spinStart() {
+    AJS.$(this.refs.spinner).show();
     AJS.$(this.refs.spinner).spin();
   }
 
@@ -169,7 +167,7 @@ class Details extends React.Component {
 
   render() {
     const { project, timestamp } = this.props.params;
-    const date = moment(+timestamp);
+    const date = moment(+ timestamp);
 
     let list;
     if (this.state.culprits === null) {
@@ -177,39 +175,50 @@ class Details extends React.Component {
     } else if (Object.keys(this.state.culprits).length === 0) {
       list = <h3>Filter matched nothing</h3>;
     } else if (Object.keys(this.state.culprits).length) {
-      list = (<List key={this.state.query.level} culprits={this.state.culprits} />);
+      list = (<List key={this.state.query.level} culprits={this.state.culprits}/>);
     }
     let overview;
     let filter;
     if (this.state.culprits !== null) {
-      overview = (<p>There are
-                      <span>&nbsp;{Object.keys(this.state.stats.urls).length}</span>
-                      &nbsp;unique URLs with a total of
-                      <span>&nbsp;{this.state.stats.count}</span>
-                      &nbsp;culprits on record.
-                      <br/>
-                      Please use the filter below to restrict the displayed culprits.
-                  </p>);
+      const totalUrls = Object.keys(this.state.stats.urls).length;
+      overview = (
+        <p className="result-count">
+          {this.state.stats.count} results from {totalUrls} URLs on record.
+          Current filter is showing {this.state.totalFiltered} from {this.state.urlsFiltered} URLs.
+        </p>
+      );
+
+      const queryStandard = this.state.query.standard instanceof Array
+        ? this.state.query.standard
+        : [this.state.query.standard];
 
       filter = (<Filter
         currentLevel={this.state.query.level}
         levels={this.state.stats.levels}
-        selectedStandards={this.state.query.standard}
+        selectedStandards={queryStandard}
         standards={this.state.stats.standards}
-        reverseDns={this.state.query.reverseDns || ''}
+        url={this.state.query.url || ''}
       />);
     }
+    const projectName = projects[project];
     return (
-        <div>
-            <h2>Details for
-                <span>&nbsp;{projects[project]}</span>
-                <span title={date.format()}>&nbsp;{date.fromNow()}</span>
-            </h2>
-            {overview}
+      <div className="aui-page-panel">
+        <div className="aui-page-panel-inner">
+          <section className="aui-page-panel-sidebar">
+            <h2>Filter</h2>
             {filter}
-            <div ref="spinner"></div>
+          </section>
+          <section className="aui-page-panel-content">
+            <h2>
+              Results for <span>{projectName}</span>
+              &nbsp;(<time title={date.format()}>{date.fromNow()}</time>)
+            </h2>
+            <div className="details-spinner" ref="spinner"></div>
+            {overview}
             {list}
+          </section>
         </div>
+      </div>
     );
   }
 }
