@@ -18,45 +18,38 @@ class Table extends React.Component {
       }
     };
 
+    const enableTooltip = (el) => AJS.$(el).tooltip({ gravity: 's' });
+
     return (
-      <table className="aui">
-        <thead>
-          <tr>
-            <th id="selector">Selector</th>
-            <th id="context">Context</th>
-            <th id="action">Action</th>
-          </tr>
-        </thead>
-        <tbody>
-          {this.props.rows.map((row) => {
-            return (
-              <tr key={row.id}>
-                <td headers="selector">
-                  <pre><code className="css" ref={highlightBlock}>{row.selector}</code></pre>
-                </td>
-                <td headers="context">
-                  {row.context
-                    ? <pre><code className="html" ref={highlightBlock}>{row.context}</code></pre>
-                    : 'n/a'}
-                </td>
-                <td className="action" headers="action">
-                  <ul className="menu">
-                    {row.help_url
-                      ? (
-                        <li>
-                          <a href={row.help_url} target="_blank">
-                            <span className="aui-icon aui-icon-small aui-iconfont-info" title="Help">Help</span>
-                          </a>
-                        </li>
-                      )
-                      : ''}
-                  </ul>
-                </td>
-              </tr>
-            );
-          })}
-        </tbody>
-      </table>
+      <ol className="context-info">
+      {this.props.rows.map((row) => {
+        let context = 'n/a';
+        if (row.context) {
+          context = <pre><code className="html" ref={highlightBlock}>{row.context}</code></pre>;
+        }
+
+        let helpUrl = null;
+        if (row.help_url) {
+          helpUrl = (<a
+            className="how-to-fix"
+            title="How to fix this error"
+            href={row.help_url}
+            ref={enableTooltip}
+            target="_blank"
+          >
+            <span className="aui-icon aui-icon-small aui-iconfont-info">How to fix</span>
+          </a>);
+        }
+
+        return (
+          <li key={row.id}>
+            {helpUrl}
+            {context}
+            <span className="selector">{row.selector}</span>
+          </li>
+        );
+      })}
+      </ol>
     );
   }
 }
