@@ -1,7 +1,5 @@
 const promisedXhr = require('./promisedXhr');
 
-module.exports = () => {
-  return promisedXhr('get', {
-    uri: '/service.json',
-  }).then((obj) => obj.body);
-};
+module.exports = () => promisedXhr('get', {
+  uri: '/service.json',
+}).then((obj) => obj.body);
